add exceptions

+++ src/main/java/io/hexlet/hexlet_spring_blog/controller/PostController.java
@@ -1,11 +1,15 @@
 package io.hexlet.hexlet_spring_blog.controller;
 
+import io.hexlet.hexlet_spring_blog.exception.ResourceNotFoundException;
 import io.hexlet.hexlet_spring_blog.model.Post;
 import io.hexlet.hexlet_spring_blog.model.PostEntity;
 import io.hexlet.hexlet_spring_blog.repository.PostRepository;
+import jakarta.validation.Valid;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -28,39 +32,41 @@ public class PostController {
     @GetMapping("/posts/{id}")
     @ResponseStatus(HttpStatus.OK)
     public PostEntity showPost(@PathVariable Long id) {
-        var post = postRepository.findById(id).get();
+        var post = postRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException(id + " Not Found"));
+
         return post;
-
-
     }
 
     @PostMapping("/posts")
     @ResponseStatus(HttpStatus.CREATED)
-    public PostEntity createPost(@RequestBody PostEntity postEntity) {
-        postRepository.save(postEntity);
-        return postEntity;
+    public PostEntity createPost(@Valid @RequestBody PostEntity postEntity) {
+        return postRepository.save(postEntity);
     }
 
     @PutMapping("/posts/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public PostEntity updatePost(@RequestBody Post data, @PathVariable Long id) {
-        var post = postRepository.findById(id).get();
+    public PostEntity updatePost(@Valid @RequestBody Post data, @PathVariable Long id) {
+        var post = postRepository.findById(id)
+                        .orElseThrow(() -> new ResourceNotFoundException(id + " Not Found"));
 
         post.setPublished(data.isPublished());
         post.setContent(data.getContent());
         post.setTitle(data.getTitle());
 
-        postRepository.save(post);
-
-        return post;
+        return postRepository.save(post);
 
     }
 
     @DeleteMapping("/posts/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deletePost(@PathVariable Long id) {
+        if(!postRepository.existsById(id)) {
+            throw new ResourceNotFoundException(id + " Not Found");
+        }
 
         postRepository.deleteById(id);
+
 
     }
 }
